Adds test so delete-tls-secret doesn't fail if secret missing

+++ scripts/delete-tls-secret.sh
@@ -3,4 +3,8 @@
 NAMESPACE="$1"
 SECRET_NAME="$2"
 
-kubectl delete secret -n "${NAMESPACE}" "${SECRET_NAME}"
+if kubectl get secret -n "${NAMESPACE}" "${SECRET_NAME}"; then
+  kubectl delete secret -n "${NAMESPACE}" "${SECRET_NAME}"
+else
+  echo "Secret ${NAMESPACE}/${SECRET_NAME} not found"
+fi
